--- main_functions/sockets.js
@@ -1,12 +1,0 @@
-exports.showMessagesHistory = function(socket)
-{
-
-}
-exports.toServerMess = function(socket)
-{
-
-}
-exports.addMessageToHistory = function(socket)
-{
-    
-}
